chore: split functions into files
